feat: request 받을 dto 추가
